pyproject.toml
[tool.poetry]
name = "1zumibotpalpha"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
chatterbot-voice = "^0.1.5"
flask = "^1.1.2"
twilio = "^6.56.0"
credentials = "^1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
